requirements.txt
aiohttp>=2.2.5
alphabet_detector
beautifulsoup4
bleach
cachetools
configparser
currencyconverter
emoji==1.7.0
feedparser
future
gitpython
gpytranslate
markdown2
multicolorcaptcha
pillow>=4.2.0
psutil
psycopg2-binary
pyYAML>=5.1.2
pyrate-limiter==2.10.0
python-telegram-bot==13.15
requests
spamwatch
speedtest-cli
sqlalchemy==1.4.11
# telethon==1.16.4
ujson
pycryptodome==3.19.0
base58
pytimeparse
SibylSystem
pydantic
redis
